third_party/WebKit/LayoutTests/inspector-protocol/accessibility/accessibility-ignoredNodes-expected.txt

Non-hidden div for comparison
non-hidden treeitem
hidden non-treeitem
 Buttons are leaf nodes
Descendant of aria-hidden node
List item also presentational
Div in list isn't presentational
Content within label refers to label container    

WebArea
  tree
    treeitem "non-hidden treeitem"
    *Ignored
      text "hidden non-treeitem"
{
  "nodeId": "<string>",
  "ignored": true,
  "ignoredReasons": [
    {
      "name": "ancestorDisallowsChild",
      "value": {
        "type": "idref",
        "relatedNodes": [
          {
            "nodeResult": "div"
          }
        ]
      }
    }
  ],
  "role": {
    "type": "internalRole",
    "value": "Ignored"
  },
  "domNode": "div"
}

WebArea
  img
    *Ignored
{
  "nodeId": "<string>",
  "ignored": true,
  "ignoredReasons": [
    {
      "name": "ancestorDisallowsChild",
      "value": {
        "type": "idref",
        "relatedNodes": [
          {
            "nodeResult": "div"
          }
        ]
      }
    }
  ],
  "role": {
    "type": "internalRole",
    "value": "Ignored"
  },
  "domNode": "svg"
}

WebArea
  button "Buttons are leaf nodes"
    *Ignored
{
  "nodeId": "<string>",
  "ignored": true,
  "ignoredReasons": [
    {
      "name": "ancestorIsLeafNode",
      "value": {
        "type": "idref",
        "relatedNodes": [
          {
            "nodeResult": "button"
          }
        ]
      }
    }
  ],
  "role": {
    "type": "internalRole",
    "value": "Ignored"
  },
  "domNode": "span"
}

WebArea
  Div
  tree
  img
  button "Buttons are leaf nodes"
  *Ignored
  text "List item also presentational"
  Div
  checkbox "Content within label refers to label container"
  Div
  Div
{
  "nodeId": "<string>",
  "ignored": true,
  "ignoredReasons": [
    {
      "name": "ariaHidden",
      "value": {
        "type": "boolean",
        "value": true
      }
    }
  ],
  "role": {
    "type": "internalRole",
    "value": "Ignored"
  },
  "domNode": "div"
}

WebArea
  Div
  tree
  img
  button "Buttons are leaf nodes"
  *Ignored
  text "List item also presentational"
  Div
  checkbox "Content within label refers to label container"
  Div
  Div
{
  "nodeId": "<string>",
  "ignored": true,
  "ignoredReasons": [
    {
      "name": "ariaHiddenRoot",
      "value": {
        "type": "idref",
        "relatedNodes": [
          {
            "nodeResult": "div"
          }
        ]
      }
    }
  ],
  "role": {
    "type": "internalRole",
    "value": "Ignored"
  },
  "domNode": "div"
}

WebArea
  Div
  tree
  img
  button "Buttons are leaf nodes"
  *Ignored
    text "List item also presentational"
    Div
  checkbox "Content within label refers to label container"
  Div
  Div
{
  "nodeId": "<string>",
  "ignored": true,
  "ignoredReasons": [
    {
      "name": "presentationalRole",
      "value": {
        "type": "boolean",
        "value": true
      }
    }
  ],
  "role": {
    "type": "internalRole",
    "value": "Ignored"
  },
  "domNode": "ol"
}

WebArea
  Div
  tree
  img
  button "Buttons are leaf nodes"
  *Ignored
    text "List item also presentational"
  Div
  checkbox "Content within label refers to label container"
  Div
  Div
{
  "nodeId": "<string>",
  "ignored": true,
  "ignoredReasons": [
    {
      "name": "inheritsPresentation",
      "value": {
        "type": "idref",
        "relatedNodes": [
          {
            "nodeResult": "ol"
          }
        ]
      }
    }
  ],
  "role": {
    "type": "internalRole",
    "value": "Ignored"
  },
  "domNode": "li"
}

WebArea
  Div
  tree
  img
  button "Buttons are leaf nodes"
  text "List item also presentational"
  *Div
    text "Div in list isn't presentational"
  checkbox "Content within label refers to label container"
  Div
  Div
{
  "nodeId": "<string>",
  "ignored": false,
  "role": {
    "type": "internalRole",
    "value": "Div"
  },
  "name": {
    "type": "computedString",
    "value": "",
    "sources": [
      {
        "type": "relatedElement",
        "attribute": "aria-labelledby"
      },
      {
        "type": "attribute",
        "attribute": "aria-label"
      },
      {
        "type": "attribute",
        "attribute": "title"
      }
    ]
  },
  "properties": [],
  "domNode": "div"
}

WebArea
  Div
  tree
  img
  button "Buttons are leaf nodes"
  text "List item also presentational"
  Div
  *Ignored
  checkbox "Content within label refers to label container"
  Div
  Div
{
  "nodeId": "<string>",
  "ignored": true,
  "ignoredReasons": [
    {
      "name": "labelFor",
      "value": {
        "type": "idref",
        "relatedNodes": [
          {
            "idref": "checkbox",
            "nodeResult": "input#checkbox"
          }
        ]
      }
    }
  ],
  "role": {
    "type": "internalRole",
    "value": "Ignored"
  },
  "domNode": "label"
}

WebArea
  Div
  tree
  img
  button "Buttons are leaf nodes"
  text "List item also presentational"
  Div
  *Ignored
  checkbox "Content within label refers to label container"
  Div
  Div
{
  "nodeId": "<string>",
  "ignored": true,
  "ignoredReasons": [
    {
      "name": "labelContainer",
      "value": {
        "type": "idref",
        "relatedNodes": [
          {
            "nodeResult": "label"
          }
        ]
      }
    },
    {
      "name": "labelFor",
      "value": {
        "type": "idref",
        "relatedNodes": [
          {
            "idref": "checkbox",
            "nodeResult": "input#checkbox"
          }
        ]
      }
    }
  ],
  "role": {
    "type": "internalRole",
    "value": "Ignored"
  },
  "domNode": "span"
}

WebArea
  Div
  tree
  img
  button "Buttons are leaf nodes"
  text "List item also presentational"
  Div
  checkbox "Content within label refers to label container"
  *Ignored
    button "aria-hidden false button"
  Div
  Div
{
  "nodeId": "<string>",
  "ignored": true,
  "ignoredReasons": [
    {
      "name": "notRendered",
      "value": {
        "type": "boolean",
        "value": true
      }
    }
  ],
  "role": {
    "type": "internalRole",
    "value": "Ignored"
  },
  "domNode": "div"
}

WebArea
  *Ignored
  button "aria-hidden false button"
{
  "nodeId": "<string>",
  "ignored": true,
  "ignoredReasons": [
    {
      "name": "notRendered",
      "value": {
        "type": "boolean",
        "value": true
      }
    }
  ],
  "role": {
    "type": "internalRole",
    "value": "Ignored"
  },
  "domNode": "span"
}

WebArea
  Div
  tree
  img
  button "Buttons are leaf nodes"
  text "List item also presentational"
  Div
  checkbox "Content within label refers to label container"
  *Ignored
  Div
  Div
{
  "nodeId": "<string>",
  "ignored": true,
  "ignoredReasons": [
    {
      "name": "probablyPresentational",
      "value": {
        "type": "boolean",
        "value": true
      }
    }
  ],
  "role": {
    "type": "internalRole",
    "value": "Ignored"
  },
  "domNode": "canvas"
}

WebArea
  Div
  tree
  img
  button "Buttons are leaf nodes"
  text "List item also presentational"
  Div
  checkbox "Content within label refers to label container"
  *Ignored
    Div
  Div
{
  "nodeId": "<string>",
  "ignored": true,
  "ignoredReasons": [
    {
      "name": "presentationalRole",
      "value": {
        "type": "boolean",
        "value": true
      }
    }
  ],
  "role": {
    "type": "internalRole",
    "value": "Ignored"
  },
  "domNode": "canvas"
}
